Refactoring & code clean up

+++ sources/Editor/EditorMenuBarWindow.h
@@ -43,7 +43,7 @@ namespace SW
         inline static const char* _fpsText = "FPS: ";
         inline static const char* _fpsTextTemplate = "FPS: 99999";
         Core::StringAtom _cachedFpsText = _fpsText;
-        const int _fpsMarginRight = 10;
+        const float _fpsMarginRight = 10.f;
 
         Core::Repeater _slowUpdater;
     };
